refactor: ConfigConverterをconfigモジュールに移動
